requirements.txt
# Core Django requirements
Django==5.2.4
djangorestframework==3.15.2

# Production database support (for SQL Server connection)
django-mssql-backend==2.8.1
pyodbc==5.1.0

# Development and testing
pytest-django==4.8.0
coverage==7.4.0